+++ pydash/collections.py
@@ -41,8 +41,6 @@ __all__ = (
     'invoke',
     'map_',
     'mapiter',
-    'max_',
-    'min_',
     'partition',
     'pluck',
     'reduce_',
@@ -554,84 +552,6 @@ def mapiter(collection, callback=None):
     """
     for result in itercallback(collection, callback):
         yield result[0]
-
-
-class _iterator_with_default(object):
-    def __init__(self, collection, default):
-        self.iter = iter(collection)
-        self.default = default
-
-    def __iter__(self):
-        return self
-
-    def next_default(self):
-        ret = self.default
-        self.default = NoValue
-        return ret
-
-    def __next__(self):
-        ret = next(self.iter, self.next_default())
-        if ret is NoValue:
-            raise StopIteration
-        return ret
-
-    next = __next__
-
-
-def max_(collection, callback=None, default=NoValue):
-    """Retrieves the maximum value of a `collection`.
-
-    Args:
-        collection (list|dict): Collection to iterate over.
-        callback (mixed, optional): Callback applied per iteration.
-        default: default value when collection is empty
-
-    Returns:
-        mixed: Maximum value.
-
-    Example:
-
-        >>> max_([1, 2, 3, 4])
-        4
-        >>> max_([{'a': 1}, {'a': 2}, {'a': 3}], 'a')
-        {'a': 3}
-        >>> max_([], default=-1)
-        -1
-
-    .. versionadded:: 1.0.0
-    """
-    if isinstance(collection, dict):
-        collection = collection.values()
-
-    return max(_iterator_with_default(collection, default),
-               key=pyd.iteratee(callback))
-
-
-def min_(collection, callback=None, default=NoValue):
-    """Retrieves the minimum value of a `collection`.
-
-    Args:
-        collection (list|dict): Collection to iterate over.
-        callback (mixed, optional): Callback applied per iteration.
-
-    Returns:
-        mixed: Minimum value.
-
-    Example:
-
-        >>> min_([1, 2, 3, 4])
-        1
-        >>> min_([{'a': 1}, {'a': 2}, {'a': 3}], 'a')
-        {'a': 1}
-        >>> min_([], default=100)
-        100
-
-    .. versionadded:: 1.0.0
-    """
-    if isinstance(collection, dict):
-        collection = collection.values()
-    return min(_iterator_with_default(collection, default),
-               key=pyd.iteratee(callback))
 
 
 def partition(collection, callback=None):