+++ tests/test_numerical.py
@@ -74,6 +74,39 @@ def test_floor(case, expected):
 
 
 @parametrize('case,expected', [
+    (([1, 2, 3],), 3),
+    (({'a': 3, 'b': 2, 'c': 1},), 3),
+])
+def test_max_(case, expected):
+    assert _.max_(*case) == expected
+
+
+@parametrize('case,expected', [
+    (([1, 2, 3],), 3),
+    (({'a': 3, 'b': 2, 'c': 1},), 3),
+    ((['anaconda', 'bison', 'camel'], lambda x: len(x)), 'anaconda'),
+    (([{'name': 'barney', 'age': 36}, {'name': 'fred', 'age': 40}], 'age',),
+     {'name': 'fred', 'age': 40}),
+    (([{'name': 'barney', 'age': 36}, {'name': 'fred', 'age': 40}],
+      lambda chr: chr['age']),
+     {'name': 'fred', 'age': 40}),
+])
+def test_max_by(case, expected):
+    assert _.max_by(*case) == expected
+
+
+@parametrize('collection,default,expected', [
+    ([], -1, -1),
+    ([1, 2, 3], -1, 3),
+    ({}, -1, -1),
+    ([], None, None),
+    ({}, None, None)
+])
+def test_max_default(collection, default, expected):
+    assert _.max_(collection, default=default) == expected
+
+
+@parametrize('case,expected', [
     (([0, 0, 0, 0, 5],), 0),
     (([0, 0, 1, 2, 5],), 1),
     (([0, 0, 1, 2],), 0.5),
@@ -81,6 +114,39 @@ def test_floor(case, expected):
 ])
 def test_median(case, expected):
     assert _.median(*case) == expected
+
+
+@parametrize('case,expected', [
+    (([1, 2, 3],), 1),
+    (({'a': 3, 'b': 2, 'c': 1},), 1),
+])
+def test_min_(case, expected):
+    assert _.min_(*case) == expected
+
+
+@parametrize('case,expected', [
+    (([1, 2, 3],), 1),
+    (({'a': 3, 'b': 2, 'c': 1},), 1),
+    ((['anaconda', 'bison', 'cat'], lambda x: len(x)), 'cat'),
+    (([{'name': 'barney', 'age': 36}, {'name': 'fred', 'age': 40}], 'age',),
+     {'name': 'barney', 'age': 36}),
+    (([{'name': 'barney', 'age': 36}, {'name': 'fred', 'age': 40}],
+      lambda chr: chr['age']),
+     {'name': 'barney', 'age': 36}),
+])
+def test_min_by(case, expected):
+    assert _.min_by(*case) == expected
+
+
+@parametrize('collection,default,expected', [
+    ([], -1, -1),
+    ([1, 2, 3], -1, 1),
+    ({}, -1, -1),
+    ([], None, None),
+    ({}, None, None)
+])
+def test_min_default(collection, default, expected):
+    assert _.min_(collection, default=default) == expected
 
 
 @parametrize('case,expected', [